refactor: use serde rename_all
Silly me, back then I didn't know that serde
had a rename_all feature to auto-correct the
field names into a certain format.

With a simple trick, we save a lot of duplicated
code.

On top of that, upgrading serde made the tests
fail, because optional fields are now treated
as error if they're null - thus we converted
some fields to `Option<T>`.